Added listing for accounts and products

+++ src/main/java/com/example/autoposterbackend/repository/PasswordResetRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PasswordResetRepository extends JpaRepository<PasswordReset, Long> {
+public interface PasswordResetRepository extends JpaRepository<PasswordReset, Integer> {
     PasswordReset findByToken(String token);
 
-    Optional<PasswordReset> findByUser_Id(Long userId);
+    Optional<PasswordReset> findByUser_Id(Integer userId);
 }
